move server code to own dir
