+++ src/client/me/mint.ts
@@ -4,7 +4,7 @@ import {
   MintApiEndpoint,
   MintApiCall,
   mintApiLocation
-} from "../../routes/me/definitions";
+} from "../../server/routes/me/definitions";
 
 import { callApi } from "../callApi";
