+++ src/server/errors/ApplicationError.ts
